Move `LabelledDividerView` into separate file
